+++ index.d.ts
@@ -1,4 +1,24 @@
 import Diagram from './@types/Diagram';
-export * from './@types/DiagramSchema';
+import useSchema from './@types/useSchema';
+import createSchema from './@types/createSchema';
+import {
+  validateSchema,
+  validatePort,
+  validateLink,
+  validateNodes,
+  validateNode,
+  validateLinks
+} from './@types/validators';
 
+export {
+  Diagram,
+  useSchema,
+  createSchema,
+  validateSchema,
+  validatePort,
+  validateLink,
+  validateNodes,
+  validateNode,
+  validateLinks
+};
 export default Diagram;