+++ src/Diagram/NodesCanvas/updateNodeCoordinates.js
@@ -1,5 +1,4 @@
 import findIndex from 'lodash.findindex';
-import cloneDeep from 'lodash.clonedeep';
 
 /**
  * Given a node id, an pair of new coordinates and the nodes array, clones the nodes array and update the node
@@ -9,11 +8,8 @@ const updateNodeCoordinates = (nodeId, coordinates, nodes) => {
   const index = findIndex(nodes, ['id', nodeId]);
 
   if (index > -1) {
-    const nextNodes = cloneDeep(nodes);
-
-    nextNodes[index].coordinates = coordinates;
-
-    return nextNodes;
+    // eslint-disable-next-line no-param-reassign
+    nodes[index].coordinates = coordinates;
   }
 
   return nodes;