+++ src/shared/Types.js
@@ -27,11 +27,19 @@ export const PortType = PropTypes.shape({
  */
 export const NodeType = PropTypes.shape({
   id: PropTypes.string.isRequired,
-  content: PropTypes.oneOfType([PropTypes.elementType, PropTypes.node]),
   coordinates: PropTypes.arrayOf(PropTypes.number).isRequired, // TODO: must be an array of 2 numbers only
+  content: PropTypes.oneOfType([PropTypes.elementType, PropTypes.node]),
   inputs: PropTypes.arrayOf(PortType),
   outputs: PropTypes.arrayOf(PortType),
   type: PropTypes.oneOf(['default']),
   render: PropTypes.elementType,
   className: PropTypes.string,
+});
+
+/**
+ * Schema
+ */
+export const SchemaType = PropTypes.shape({
+  nodes: PropTypes.arrayOf(NodeType).isRequired,
+  links: PropTypes.arrayOf(LinkType),
 });